로그인, 리스트

+++ src/commons/pay/payment.queries.ts
@@ -1,7 +1,7 @@
 import { gql } from "@apollo/client";
 
 export const CREATE_PAYMENT = gql`
-  mutation createPayment($impUid: String!, $amount: Float!) {
+  mutation createPayment($impUid: String!, $amount: Int!) {
     createPayment(impUid: $impUid, amount: $amount) {
       # paymentDate
       # paymentAmount
